requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
boto3>=1.28.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pypdf>=3.0.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
